Main code framework and minor additions to sensors
The .ino takes the rover to its landed state. Currently there are several print statements for debugging purposes which will eventually need to be removed. THIS CODE IS STILL UNTESTED. 
Each individual sensor was given a method to set its registers for sleep and normal operation. I also added a constant so that the barometer outputs in feet instead of meters.