Add files via upload
The .ino takes the rover to its landed state. Currently there are several print statements for debugging purposes which will eventually need to be removed. THIS CODE IS STILL UNTESTED. 
Each individual sensor was given a method to set its registers for sleep and normal operation. The IMU also now has methods to set these states for all 4 sensors. I also added a constant so that the barometer outputs in feet instead of meters.